refactor(runtime): extract builtin_tool_definitions into definitions module
The ~1.1k-line builtin_tool_definitions() function — every native
tool's JSON schema — moves into tool_runner/definitions.rs along
with its ALWAYS_NATIVE_TOOLS constant (#3044) and select_native_tools
helper. mod.rs re-exports all three with `pub use` so external
consumers (librefang_kernel::kernel::mod, the test suite, the
auto_dream module) keep resolving the same paths
(`librefang_runtime::tool_runner::builtin_tool_definitions`, etc.).

definitions.rs reaches SUPPORTED_AUDIO_EXTS_DOC via `super::media`
— sibling access is fine because the const is pub(super) inside
tool_runner and definitions.rs is also a child of tool_runner.

While here, replace the stale `// Inter-agent tools` section header
above require_kernel with a one-line doc comment, since the
inter-agent tools themselves no longer live next to it.

mod.rs: 9,171 -> 7,976 LOC (-1195). cargo check + cargo clippy --lib
both clean. The single largest extraction in this issue so far.

Refs #3710
